Added code to clean up bulk_logs dir
Added INI path for bulk_logs and temp.  In bulk processing the gui app
will be responsible for creating the bulk_logs dir and putting in the
logs to process.  When iislp is done it will remove that dir and its
contents and that will signal to the gui app that it is done.  That code
is in here now to remove the dir after the "Cleaning up..." log item.