Updated threading and runlog list updating
Added handling for pausing/resuming the runlog thread.  On launch it
reads the file and updates the listbox in the view.  Then it pauses.  It
resumes when the bulk process button is pressed and stops again when it
reads in the line of the log file that contains "Killing all processes"
which signals the end of the log output for the bulk processing.