More work on bulk_process method in the controller
This method handles all the steps that need to be taken to perform a
bulk process of any number of logs.  It is the point where iislp and
iislp_gui meet and interact.  This method decides how to control the
iislp.  Made an algorithm that gets me all of the unprocessed logs in 40
log batches with the last batch possibly being the remainder of logs.
All the batches are saved in a dictionary.